starting the parallel processing

+++ src/main/java/com/example/services/CustomerServices.java
@@ -29,7 +29,7 @@ public interface CustomerServices {
      * @param longitude
      * @return
      */
-    public Address lookupAddress(long latitude, long longitude);
+    public Address lookupAddress(double latitude, double longitude);
 
     /**
      *
